samples/ml/ml_jobs/single-node/xgb-loan-apps/requirements.txt
snowflake-ml-python>=1.7.4
jupyter
xgboost